Squashed commit of the following:
commit 82650815327dbce9e8d4fed1ac6d140186ace981
Author: annisafrl <annisafirlia1@gmail.com>
Date:   Thu Jun 22 17:11:36 2023 +0700

    fp-15 add sortbyabjad

commit 634c2c7f10d1148eb87cd6e789a8d4a8f2d18a34
Author: annisafrl <annisafirlia1@gmail.com>
Date:   Thu Jun 22 14:04:55 2023 +0700

    update app.js

commit 894c2dc67f4fca9caa15840f25308b6a09ae059d
Author: annisafrl <annisafirlia1@gmail.com>
Date:   Thu Jun 22 13:59:45 2023 +0700

    fp-15 update app.js

commit 22fb2b6208fd397e7944af3610bfbd796e48e221
Author: annisafrl <annisafirlia1@gmail.com>
Date:   Thu Jun 22 13:58:55 2023 +0700

    fp-15 update navbar before login and after login

commit 6db65fb1399e0507bbc6c1372d8d3ae0519c13de
Author: annisafrl <annisafirlia1@gmail.com>
Date:   Thu Jun 22 10:15:11 2023 +0700

    fp-15 update border cartmodal

commit 0b403dd6b5a4363973cf29f7df8d909aaa09188f
Author: annisafrl <annisafirlia1@gmail.com>
Date:   Wed Jun 21 16:00:10 2023 +0700

    fp-15 update empty cart handler

commit bc10115a14080702c4c85ca42179a2e59b686cdc
Author: annisafrl <annisafirlia1@gmail.com>
Date:   Wed Jun 21 10:33:42 2023 +0700

    fp-15 add cart fitur

+++ projects/client/src/pages/Cart/Cart.jsx
@@ -4,26 +4,60 @@ import {
   decreaseQuantity,
   increaseQuantity,
 } from "../../features/cart/cartSlice";
+import Swal from "sweetalert2";
+import emptyCart from "../../img/EmptyCart_3-Copy-removebg-preview.png";
 
 function Cart() {
   const dispatch = useDispatch();
   const cartItems = useSelector((state) => state.cart.items);
 
   const handleRemoveItem = (id) => {
-    dispatch(removeItem(id));
+    Swal.fire({
+      title: "Are you sure?",
+      text: "This item will be removed from your cart.",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonText: "Yes, remove it!",
+      cancelButtonText: "Cancel",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        dispatch(removeItem(id));
+        Swal.fire(
+          "Removed!",
+          "The item has been removed from your cart.",
+          "success"
+        );
+      }
+    });
   };
 
   const handleDecreaseQuantity = (id) => {
     const item = cartItems.find((item) => item.id_product === id);
     if (item.quantity === 1) {
-      handleRemoveItem(id);
+      Swal.fire({
+        title: "Are you sure?",
+        text: "This item will be removed from your cart.",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonText: "Yes, remove it!",
+        cancelButtonText: "Cancel",
+      }).then((result) => {
+        if (result.isConfirmed) {
+          dispatch(removeItem(id));
+          Swal.fire(
+            "Removed!",
+            "The item has been removed from your cart.",
+            "success"
+          );
+        }
+      });
     } else {
       dispatch(decreaseQuantity(id));
     }
   };
 
   const handleIncreaseQuantity = (id) => {
-    alert(id);
+    // alert(id);
     dispatch(increaseQuantity(id));
   };
 
@@ -33,11 +67,15 @@ function Cart() {
   );
 
   return (
-    <div className="flex flex-col h-10 mx-auto max-w-2xl my-20">
-      <p className="text-lg font-bold text-white text-center">Shopping Cart</p>
-
+    <div className="flex flex-col h-30 mx-auto max-w-l mt-20">
       {cartItems.length === 0 ? (
-        <p className="text-lg text-white">Your cart is empty.</p>
+        <div className="flex flex-col items-center justify-center">
+          <img src={emptyCart} alt="Empty Cart" className="w-40 h-30" />
+          <p className="text-lg text-red-600 mt-4">Your cart is empty.</p>
+          <p className="text-sm text-gray-500 mt-2">
+            It looks like you haven't added any products to your cart.
+          </p>
+        </div>
       ) : (
         <ul className="divide-y divide-[#EDA415]">
           {cartItems.map((item) => (
@@ -45,12 +83,9 @@ function Cart() {
               <div className="ml-6 flex-1 flex flex-col justify-between">
                 <div className="flex">
                   <div className="flex-1">
-                    <h2 className="text-lg font-medium text-white">
+                    <h2 className="text-lg font-medium text-[#EDA415]">
                       {item.product_name}
                     </h2>
-                    <p className="mt-1 text-sm text-white-500">
-                      {item.product_description}
-                    </p>
                   </div>
                   <div className="ml-4 flex-shrink-0 flow-root">
                     <button
@@ -65,7 +100,7 @@ function Cart() {
                 <div className="flex-1 flex flex-col justify-between">
                   <div className="flex">
                     <div className="flex-shrink-0">
-                      <p className="text-lg font-medium text-white">
+                      <p className="text-l font-medium">
                         {item.product_price.toLocaleString("id-ID", {
                           style: "currency",
                           currency: "IDR",
@@ -96,8 +131,8 @@ function Cart() {
             </li>
           ))}
           <div class="mt-6 flex items-center justify-between">
-            <p class="text-sm font-medium text-white">Total</p>
-            <p class="text-2xl font-semibold text-white">
+            <p class="text-sm font-medium text-[#EDA415]">Total</p>
+            <p class="text-l font-semibold">
               {" "}
               {totalPrice.toLocaleString("id-ID", {
                 style: "currency",
